Fixed bug where linking directly to a note would not work
Previously when a url like /issues/1234#note-4 was visited it would be immediately replaced with /issues/1234?tab=history_all, and the page would not focus on the note.

This change checks if there is a note in the url fragment and if there is, displays the 'all' tab, and scrolls to/ highlights the note.

Also refactored the initTabs function to make it easier to add more tabs and to select tabs by id rather than id+journal class

Change-Id: I1e1385f66c3235c046299dc93d7a9dd83a9d1731